refactor: structure webhook endpoints for future expansion
