+++ vitest.config.ts
@@ -23,9 +23,9 @@ export default defineConfig({
         "convex/schema.ts",
         "convex/users/schema.ts",
         "convex/http.ts",
+        "convex/webhooks/**",
         "convex/**/mutations.ts",
         "convex/**/webhooks.ts",
-        "convex/**/clerkWebhook.ts",
         "**/*.d.ts",
         "**/*.test.ts",
       ],
